no rm -f *$OBJSUF if $OBJSUF is empty
git-svn-id: https://svn.code.sf.net/p/zzjson/code/trunk@16 fca5a30d-054f-0410-8496-fa6b1410ff5f

+++ buildtest.sh
@@ -17,4 +17,4 @@ echo
 echo "*** test program, size optimized, stripped"
 execute $CC -o ${EXEBASE}_s${EXESUF} zzjson_*.c $COMMON $OPTSIZ $STRIP
 
-rm -f *${OBJSUF}
+test -n "$OBJSUF" && rm -f *${OBJSUF}
